fix(测试计划): 修复测试计划批量执行时出现的问题
测试计划批量执行（尤其是串行）时，由于每个测试计划执行是需要记录执行产生的报告并入库，在测试计划数量较多时会出现由于事务导致的数据一致性问题：测试计划批量执行方法还在进行，测试计划报告数据还未入库，但是先执行的测试计划已经执行结束并通过kafka发送回消息。此时由于数据并未入库，也就无法更改测试计划报告状态，导致测试计划报告状态停留在了入库时的running
